Version 31 de Diciembre 2020

+++ src/main/java/com/myfactory/SBootWebProject/repository/EmpleadoJPARepository.java
@@ -1,11 +1,11 @@
 package com.myfactory.SBootWebProject.repository;
 
-
 import org.springframework.data.repository.CrudRepository;
+
 import com.myfactory.SBootWebProject.model.Empleado;
 
-
-public interface EmpleadoJPARepository extends CrudRepository<Empleado, Integer> {
+public interface EmpleadoJPARepository extends CrudRepository<Empleado, Long> {
+	
 		//  @Query("select cli from Cliente cli where cli.apellidos = :apellidos")
 		 // Iterable <Empleado>  findByApellidos(@Param("apellidos") String apellidos)  ;
